+++ BTCPayServer.Plugins.LNbank/Services/Wallets/WalletsQuery.cs
@@ -1,3 +1,6 @@
+using System.Linq;
+using BTCPayServer.Plugins.LNbank.Data.Models;
+
 namespace BTCPayServer.Plugins.LNbank.Services.Wallets;
 
 public class WalletsQuery
@@ -8,6 +11,10 @@ public class WalletsQuery
     public bool IncludeTransactions { get; set; }
     public bool IncludeAccessKeys { get; set; }
     public bool IncludeUser { get; set; }
-
     public bool IsServerAdmin { get; set; }
+
+    public bool HasAdminAccess(Wallet wallet)
+    {
+        return IsServerAdmin || (UserId != null && UserId.Contains(wallet.UserId));
+    }
 }